Finish extracting google-specific stuff.
