+++ src/realtime/index.ts
@@ -1,5 +1,3 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
-//
 export * from './realtime';
-export * from './plugin';
